+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.3 [UNRELEASED]
 - Fix invalid import in test helper
+- Fix race condition when using nested changesets
 
 ## 1.0.2 [2025-09-20]
 - Fix invalid require path preventing gem from loading